ISearchResponse<T>  deserialization prototype
This commit is a minimum viable prototype of using utf8json serializer to
deserializer an ISearchResponse<T>. Utf8Json unit test has been used to
test deserialization.

Notable points:

1. Hit<T> requires a custom formatter to be resolved at the
IJsonFormatterResolver level because it contains a generic type property
whose formatter, SourceFormatter<T> cannot be resolved using JsonFormatterAttribute. If
this can be resolved, then it would be possible to attribute Hit<T> with JsonFormatterAttribute
and typeof(HitFormatter<>), and have typeof(SourceFormatter<>) resolved for the _source field.

  For now instead, initialize the SourceFormatter<T> inside the HitFormatter<T> ctor.
2. Implementation does not handle different field casings
3. HitFormatter<T> avoids allocations for property names by using AutomataDictionary. This dictionary
lives outside of the generic HitFormatter<T> to avoid creating an instance for each T.
4. Both JsonReader and JsonWriter are structs passed by ref, so cannot be captured inside of local
functions or lambda expression bodies. Would need to be passed as a ref parameter. An example
is JoinFieldFormatter's Serialize method.