Fix property mapping serialization
This commit fixes the serialization of property mappings.
IProperties is a dictionary of PropertyName -> IProperty,
where IProperty is a concrete instance of a derived property type
that implements an interface derived from IProperty. In such
a setup, each IProperty should be serialized with a generic formatter
typed to the derived interface type to ensure all properties marked with
DataMemberAttribute are serialized.